requirements.txt
alembic==1.14.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.10.4
asyncpg==0.30.0
bcrypt==4.0.1
boto3==1.35.97
botocore==1.35.97
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.115.6
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
passlib==1.7.4
Pillow==11.1.0
pyasn1==0.6.1
pycparser==2.23
PyPDF2==3.0.1
pymupdf==1.23.26
reportlab==4.2.5
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dateutil==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
python-telegram-bot==21.10
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.34.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
pytz>=2024.1
